+++ packages/client/src/components/dispatch/map/calls/RenderActiveCalls.tsx
@@ -112,7 +112,6 @@ export function RenderActiveCalls() {
         );
       })}
 
-      {/* <ActiveMapUnits /> */}
       <ActiveMapCalls
         openItems={openItems}
         setOpenItems={setOpenItems}
